데이터 등록 화면 파일 관리 테이블 구현 중

+++ SmartKMS/src/main/java/com/innotree/smartkms/dataanalysis/controller/DataManagerController.java
@@ -57,7 +57,7 @@ public class DataManagerController {
 	
 	@GetMapping("/startimport")
 	@ResponseBody
-	public synchronized Map<String, Object> startImport(@RequestParam("id") int id, 
+	public synchronized Map<String, Object> startImport(@RequestParam("id") long id, 
 													   @RequestParam("fileName") String fileName,
 													   @RequestParam("indexName") String indexName,
 													   @RequestParam("type") String type) {
